MVP portfolio masonry grid without any sorting functionality
