+++ src/pages/PortfolioPage/PortfolioPage.css
@@ -1,0 +1,10 @@
+.page {
+  padding: 20px;
+}
+
+.portfolioGrid  h1 {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+}
