--- src/pages/portfolio/Portfolio.css
@@ -1,8 +1,0 @@
-.text-center {
-  text-align: center;
-}
-
-.page {
-  padding: 20px;
-  align-items: center;
-}
